externalize function to get snapshot filename
